Clarifying the license declaration.
The license in the README should properly acknowledge Zach's
contributions. Zach had forgotten to update his license dates in his own
project, but international law automatically applies copyright when
something is recorded in a fixed form. Given that 0.3.0 (on which rupy
is based) was released in 2010, the range is properly 2008 - 2010.